requirements.txt
pip==24.3.1
Flask==3.1.0
solana==0.36.2
flask-cors==5.0.0
pynacl=1.5.0
solders=0.23.0